requirements.txt
PyYAML>=6.0
pathlib2>=2.3.0